+++ tests/projects/embed/iverilog/hello_vcd/src/main.v
@@ -1,6 +1,8 @@
 module hello;
   initial begin
     $display("hello world!");
+    $dumpfile("hello.vcd");
+    $dumpvars(0, hello);
     $finish ;
   end
 endmodule
