+++ tests/projects/embed/iverilog/hello_vcd/xmake.lua
@@ -1,0 +1,5 @@
+add_requires("iverilog")
+target("hello")
+    add_rules("iverilog.binary")
+    set_toolchains("@iverilog")
+    add_files("src/*.v")
