+++ tests/projects/embed/verilator/hello/src/main.v
@@ -1,7 +1,6 @@
 module hello;
   initial begin
     $display("hello world!");
-    $dumpvars(0, hello);
     $finish ;
   end
 endmodule
