+++ tests/projects/embed/verilator/hello/src/sim_main.cpp
@@ -1,26 +1,10 @@
 #include "hello.h"
 #include "verilated.h"
-#include "verilated_vcd_c.h"
-
-double sc_time_stamp() {
-    return 0;
-}
 
 int main(int argc, char** argv) {
-    char const* vcdfile = NULL;
-    if (argc == 2) {
-        vcdfile = argv[1];
-    } else {
-        printf("xmake run hello [vcdfile]\n");
-        return -1;
-    }
     VerilatedContext* contextp = new VerilatedContext;
     contextp->commandArgs(argc, argv);
     hello* top = new hello{contextp};
-    VerilatedVcdC* tfp = new VerilatedVcdC;
-    top->trace(tfp, 99);
-    Verilated::traceEverOn(true);
-    tfp->open(vcdfile);
     while (!contextp->gotFinish()) { top->eval(); }
     delete top;
     delete contextp;
