+++ tests/projects/embed/verilator/hello_vcd/src/main.v
@@ -1,6 +1,7 @@
 module hello;
   initial begin
     $display("hello world!");
+    $dumpvars(0, hello);
     $finish ;
   end
 endmodule
