+++ tests/projects/embed/verilator/hello_vcd/xmake.lua
@@ -4,3 +4,4 @@ target("hello")
     set_toolchains("@verilator")
     add_files("src/*.v")
     add_files("src/*.cpp")
+    add_values("verilator.flags", "--trace", "--timing")
